# contracts/generated/api-v2.openapi.json
{
  "components": {
    "schemas": {
...
      "ArtifactCreate": {
        "additionalProperties": false,
        "properties": {
          "content": {
            "description": "At most 2 MiB of UTF-8, including JSON whitespace; the original bytes are preserved.",
            "maxLength": 2097152,
            "minLength": 1,
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/ResearchArtifactKind"
          },
          "project_id": {
            "$ref": "#/components/schemas/Id"
          },
          "schema_version": {
            "$ref": "#/components/schemas/SchemaV1"
          }
        },
        "required": [
          "schema_version",
          "project_id",
          "kind",
          "content"
        ],
        "type": "object"
      },
...
      "Id": {
        "format": "uuid",
        "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-7[0-9a-fA-F]{3}-[89aAbB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$",
        "type": "string"
      },
...
      "ResearchArtifactKind": {
        "enum": [
          "CODE",
          "PARAMETERS",
          "REPORT"
        ],
        "type": "string"
      },
...
      "SchemaV1": {
        "enum": [
          1
        ],
        "maximum": 1,
        "minimum": 1,
        "type": "integer"
      },
...
    },
    "securitySchemes": {
      "BrowserSession": {
        "description": "Native private cookie; browser writes require exact same-origin Origin. Explicit loopback development uses quazonai-dev.",
        "in": "cookie",
        "name": "__Host-quazonai",
        "type": "apiKey"
      },
      "MachineBearer": {
        "bearerFormat": "qz2.UUIDv7.opaque-capability",
        "description": "Opaque native capability; only project/run/downstream-scoped server records confer authority. Never combine with browser Cookie.",
        "scheme": "bearer",
        "type": "http"
      },
      "OperatorCommandGrant": {
        "description": "One-time TOTP-verified CLI grant bound to this credential, exact operation, target and full nonsecret request. No Agent/automation grant issuance.",
        "in": "header",
        "name": "X-Operator-Grant",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "",
    "license": {
      "name": ""
    },
    "title": "QuaZonai HTTP API",
    "version": "2.0.0-dev.1"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v2/artifacts": {
      "get": {
        "operationId": "list",
        "parameters": [
          {
            "in": "query",
            "name": "project_id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Id"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "format": "int32",
              "maximum": 100,
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ArtifactView"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": ""
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": ""
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Shared native machine attempt limit or bounded crypto capacity; respect Retry-After.",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "BrowserSession": []
          },
          {
            "MachineBearer": []
          }
        ],
        "tags": [
          "Artifacts"
        ]
      },
      "post": {
        "operationId": "create",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArtifactCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResult_ArtifactView"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": ""
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": ""
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": ""
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Shared native machine attempt limit or bounded crypto capacity; respect Retry-After.",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "BrowserSession": []
          },
          {
            "MachineBearer": []
          }
        ],
        "tags": [
          "Artifacts"
        ]
      }
    },
    "/api/v2/artifacts/{id}": {
      "get": {
        "operationId": "get",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactView"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": ""
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Shared native machine attempt limit or bounded crypto capacity; respect Retry-After.",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "BrowserSession": []
          },
          {
            "MachineBearer": []
          }
        ],
        "tags": [
          "Artifacts"
        ]
      }
    },
    "/api/v2/artifacts/{id}/content": {
      "get": {
        "operationId": "content",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/octet-stream": {
                "schema": {
                  "description": "OpenAPI-only binary shape; the actual handler streams the original bytes.",
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Attachment with original native bytes; never rendered inline."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": ""
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Shared native machine attempt limit or bounded crypto capacity; respect Retry-After.",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "BrowserSession": []
          },
          {
            "MachineBearer": []
          }
        ],
        "tags": [
          "Artifacts"
        ]
      }
    },
    "/api/v2/auth/devices": {
      "get": {
        "operationId": "devices",
        "parameters": [
          {
            "description": "Opaque previous-page cursor",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceList"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "BrowserSession": []
          }
        ],
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v2/auth/devices/{id}": {
      "delete": {
        "operationId": "revoke_device",
        "parameters": [
          {
            "description": "Trusted device identity",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "BrowserSession": []
          }
        ],
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v2/auth/login": {
      "post": {
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrowserSession"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": ""
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": ""
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": ""
          }
        },
        "security": [],
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v2/auth/logout": {
      "post": {
        "operationId": "logout",
        "responses": {
          "204": {
            "description": ""
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "BrowserSession": []
          }
        ],
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v2/auth/machine": {
      "get": {
        "operationId": "machine_session",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineSessionView"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": ""
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Shared native machine attempt limit or bounded crypto capacity; respect Retry-After.",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "MachineBearer": []
          }
        ],
        "tags": [
          "Machine identity"
        ]
      }
    },
...
    "/api/v2/auth/session": {
      "get": {
        "operationId": "session_status",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrowserSession"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "BrowserSession": []
          }
        ],
        "tags": [
          "Authentication"
        ]
      }
    },
...
  },
...
}